VirtualFileSystem: Implement write/append support for OpenFile()
In order to properly mock this function, use temporary files to
allow fopen() to open stdio descriptors to them in write mode.

This is achieved by adding a new 'writable_path' string path
field in VirtualFileSystem::Entry. The the path is not empty,
then it will point to a temporary file used to back the content
of the file.

Note that its location is indendent from the actual file path used
in the virtual filesystem.

The temporary file is removed automatically once the
VirtualFileSystem instance is closed, or if RemoveFile() is
used on the corresponding entry.