Print a table of running commands in interactive terminals.
Ninja will now print a table of up to 4 running commands below
the status line, when in interactive terminals (nothing in
changed for non-interactive ones).

The size of the table can be changed by setting
`NINJA_STATUS_MAX_COMMANDS=<count>` in the environment,
and using a value of 0, or a negative one, disables
the feature.

While environment variables are generally frawned upon,
the value being used here does not change Ninja's behavior
per se (e.g. changing it does not require a rebuild, or
changes the output of tools), and is the best compromise
to allow users to change it without requiring a dedicated
config file and parser.

This also changes the default refresh update timeout to 100ms
for more dynamic updates, except if the feature is disabled
(in which case the default stays at 1s for time-sensitive
status updates).