includes_normalize: Partial fix for long path support.
Do not use fixed-size buffers in order to better support
long file paths. This does *not* get rid of the failure
described at [1], because GetFullPathNameA() will fail
with long paths, even when the feature is enabled in
the application's manifest.

Insteead, using GetFullPathNameW() is required to completely
fix this issue.

[1] https://github.com/ninja-build/ninja/issues/2442#issuecomment-2585728337