subprocess-posix.c: Use FdWaiter class to wait for async events.
Introduce FdWaiter class to implement waiting for async read
events on a set of file descriptors using either pselect() or ppoll(),
then use it on SubprocessSet::DoWork().

This refactor removes some duplication and helps clarify what is
going on in Subprocess::DoWork(). Moreover, it prepares a future
change that will be necessary to split stdout and stderr buffering
for each non-console process.