Implement stdout/stderr split for subprocess-win32.cc
Ensure that each Subprocess instance stores separate buffers
for the stdout and stderr outputs of non-console sub-processses.

Note that Subprocess::GetOutput() still retrieves the combined
stdout + stderr output to avoid changing anything else in the
Ninja code base, or breaking any existing developer workflows
which may rely on that. Though the fact of using two pipes
instead of one may alter the exact sequence of bytes in the
output when the child process writes to both outputs at
the same time.