Properly separate stdout and stderr for each subprocess.
Remove the Subprocess::GetOutput() method and corresponding
combined_output_ field. Instead callers should now use
Subprocess::GetStdout() and Subprocess::GetStderr() as appropriate.

In practice, the major change in behavior is that Ninja will now
always output a command's entire stdout followed by its entire stderr
output on completion. Before that, the two were mixed in unpredictable
ways when the process wrote to both streams at the same time.

This operation is performed as the last commit of a series to
make it easy to revert, in case this breaks some developer workflows
(which would be surprising, as how the two streams are mixed is
not in general unpredictable).